refactor: extract BlockKeywordMatcher; address remaining sonar nitpicks
SonarCloud:
- java:S1200 (FormatProvider 21/20 dep): вынес ScopePair + 5 SCOPE-констант +
  SCOPE_BY_CLOSER_TYPE + findMatchingOpener в новый package-private класс
  BlockKeywordMatcher. FormatProvider теперь зовёт
  BlockKeywordMatcher.findMatchingOpener(...) и не зависит от ScopePair напрямую.
- java:S1213 (×6): после извлечения static-полей в отдельный класс порядка
  объявлений в FormatProvider больше не нарушает соглашение.
- java:S1941: targetIndent объявляется после baseEdits — ближе к месту
  использования, и не считается, если baseEdits пустой.

Тесты FormatProviderTest: 20/20 зелёные.
